Restructure routes, pages and layouts
