+++ core/app/(auth)/(forms)/register/page.tsx
