+++ core/app/(auth)/email-verification/layout.tsx
