+++ core/app/(auth)/error/page.tsx
