--- core/app/(landing)/(auth)/(forms)/reset/page.tsx
@@ -1,9 +1,0 @@
-import { ResetPasswordForm } from "@/components/authentication/ResetPasswordForm";
-
-const ResetPasswordPage = () => {
-    return (
-        <ResetPasswordForm />
-    );
-};
-
-export default ResetPasswordPage;
