--- core/app/(landing)/(auth)/email-verification/page.tsx
@@ -1,9 +1,0 @@
-import { EmailVerificationForm } from "@/components/authentication/EmailVerificationForm";
-
-const EmailVerificationPage = () => {
-    return (
-        <EmailVerificationForm/>
-    );
-}
-
-export default EmailVerificationPage;
